+++ twister2/data/src/main/java/edu/iu/dsc/tws/data/api/splits/GenericCSVInputSplit.java
@@ -9,9 +9,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package edu.iu.dsc.tws.executor.threading;
+package edu.iu.dsc.tws.data.api.splits;
 
-import edu.iu.dsc.tws.executor.api.IExecutor;
-
-public abstract class AbstractExecutor implements IExecutor {
+public class GenericCSVInputSplit {
 }
