+++ twister2/data/src/main/java/edu/iu/dsc/tws/dataset/ExecutionSchedule.java
@@ -1,0 +1,31 @@
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+package edu.iu.dsc.tws.dataset;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+
+/**
+ * Represent the schedule of execution units
+ */
+public class ExecutionSchedule {
+  /**
+   * The id of this worker
+   */
+  private int thisWorker;
+
+  /**
+   * The worker id to executor mapping
+   */
+  private Map<Integer, Set<Integer>> schedule = new HashMap<>();
+}
