+++ twister2/executor/src/java/edu/iu/dsc/tws/executor/core/Runtime.java
@@ -13,18 +13,42 @@ package edu.iu.dsc.tws.executor.core;
 
 import edu.iu.dsc.tws.common.config.Config;
 import edu.iu.dsc.tws.common.config.Context;
+import edu.iu.dsc.tws.comms.api.TWSChannel;
 import edu.iu.dsc.tws.data.api.InputFormat;
 import edu.iu.dsc.tws.data.fs.Path;
 import edu.iu.dsc.tws.dataset.DataSource;
+import edu.iu.dsc.tws.executor.api.ExecutionPlan;
 import edu.iu.dsc.tws.task.api.TaskContext;
 
 /**
  * Captures the runtime information about the system.
  */
 public class Runtime {
+  /**
+   * Name of the job
+   */
   private String jobName;
 
+  /**
+   * The job directory
+   */
   private Path parentpath;
+
+  /**
+   * Execution plan
+   */
+  private ExecutionPlan plan;
+
+  /**
+   * The communication channel
+   */
+  private TWSChannel channel;
+
+  public Runtime(String jName, ExecutionPlan execPlan, TWSChannel ch) {
+    this.jobName = jName;
+    this.plan = execPlan;
+    this.channel = ch;
+  }
 
   public String getJobName() {
     return jobName;
@@ -46,7 +70,7 @@ public class Runtime {
     this.parentpath = parentpath;
   }
 
-  public <T> DataSource<T> createInput(TaskContext context, InputFormat<T, ?> input) {
-    return null;
+  public <T> DataSource<T> createInput(Config cfg, TaskContext context, InputFormat<T, ?> input) {
+    return new DataSource<>(cfg, input, context.getParallelism());
   }
 }
